bio: fix a bug of partial accesses.
In case of pre partition access, the value of 'block' variable
has been added to valid block counts, but since the value is modified
to read as mush as a size of native block size before,
the value must be aligned to a size of native block size in here and
thus the values should be added to a size of native block size.

And even with non-alignment of an end of the requested range,
new read has been just returned. In this case, one partial access
should be done to handle the partial amount.

Change-Id: Ie4cefa5ffddaa1a605aa30b66a678a820a065098
Signed-off-by: Kiwoong Kim <kwmad.kim@samsung.com>