[850] PLATFORM: save dump_en firstly and clear register
DUMP_EN affects several solutions. Becuase DUMP_EN is related with cache operation.
If the system is stuck while running on lk with cache turned on,
the next LK should not use the cache. because the cache is stopped by DUMP_EN.
So we fix that we save DUMP_EN register to variable and DUMP_EN register clear.

Change-Id: I97aa1d8531277baa2298b89d6f2114557f1c5aa9
Signed-off-by: Donghyeok Choe <d7271.choe@samsung.com>